requirements.txt
Django==1.11
requests==2.24.0
django_rq==1.3.1
djangorestframework==3.11.1
psycopg2==2.8.6
